Move actual build to 'before_script' section
A build failure in 'before_script' marks the whole job as 'errored'
and prevents from entering the following sections. On the other
hand, errors within 'script' mark the job as 'failed', but will not
cancel it right away. The point of this commit is to provide a means
of distinguishing build/test failures and to abort the execution of
test scripts if there are no binaries to link to.