Fix std::vector's storage handling, void reserve() method
See d59c5d4. It turns out that reserve() does allocate memory and allows
further calls to operator[], but the reserved *capacity* does not matter
when resize() is called, that is:

  std::vector<int> v; // size = capacity = 0
  v.reserve(1); // next instruction would fail without this
  v[0] = 1;
  v.resize(1); // same as v.resize(1, 0)
  std::cout < v[0];

yields '0' since resize() sees a zero-sized vector and overwrites the
'unmanaged' elements (also, at() would throw an exception). This, in
turn, does work when push_back() is used instead.